refactor(service-overview-small): removed service item small
