+++ components/services-overview-small/services-overview-small.stories.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import ServicesOverviewSmall from '../services-overview-small/services-overview-small';
+import ServicesOverviewSmall from './services-overview-small';
 import Data from '../../data/current/home.json';
 
-storiesOf('Service item small', module)
-	.add('service item small', () => (
+storiesOf('Service overview small', module)
+	.add('service overview small', () => (
 		<ServicesOverviewSmall title={Data.servicesItemTitle}  services={Data.serviceItems} />
 	));
