+++ styles/index.less
@@ -36,6 +36,7 @@
 @import (once) '../components/logo/logo';
 @import (once) '../components/parallax/parallax';
 @import (once) '../components/services-item/services-item';
+@import (once) '../components/services-overview-small/services-overview-small';
 @import (once) '../components/services-overview/services-overview';
 @import (once) '../components/services-overview/services-overview-shapes';
 @import (once) '../components/shapes/shapes';
@@ -44,7 +45,6 @@
 @import (once) '../components/social-media/social-media';
 @import (once) '../components/footer/footer';
 @import (once) '../components/footer-locations/footer-locations';
-@import (once) '../components/services-item-small/services-item-small';
 @import (once) '../components/events-extract/events-extract';
 @import (once) '../components/feeds-block/feeds-block';
 @import (once) '../components/image-feed/image-feed';
